requirements.txt
scrapy
beautifulsoup4
requests
peewee
python-dotenv